Recoded DAC module in more correct form

+++ dac/driver/LPC177x_8x_dac.h
@@ -1,11 +1,12 @@
 /****************************************************************************
 LPC177x_8x_dac.h basic definitions
 ****************************************************************************/
+#include "LPC177x_8x.h"
 
 #define DACR_BIAS16 	16
 
-void InitDAC (unsigned int val);
-void SetDAC (unsigned int val);
+void	InitDAC	(uint16_t val);
+void	SetDAC	(uint16_t val);
 
 /*****************************************************************************
 End Of File
